backend/requirements.txt
fastapi
sqlalchemy
asyncpg
python-jose
passlib[bcrypt]==1.7.4
pydantic
python-dotenv
uvicorn
psycopg2-binary
bcrypt==4.0.1
